Cast target columns to float after fillna
Pandas reads all-NULL SQLite columns as object dtype; fillna(0) alone
doesn't coerce the type, causing torch.as_tensor() to fail at inference

+++ forecast.py
@@ -55,7 +55,7 @@ def _prepare_hourly(update: bool = True) -> pd.DataFrame:
         update=update,
     )
     data['hour'] = data['datetime'].dt.hour
-    data[_hcfg.targets] = data[_hcfg.targets].fillna(0)
+    data[_hcfg.targets] = data[_hcfg.targets].fillna(0).astype(float)
     _warn_missing(data, _hcfg.real_unknown, 'hourly')
     data[_hcfg.real_unknown] = data[_hcfg.real_unknown].interpolate(method='linear', limit=2).ffill().bfill()
     data[_hcfg.real_known]   = data[_hcfg.real_known].interpolate(method='linear')
@@ -127,7 +127,7 @@ def _prepare_daily(update: bool = True) -> pd.DataFrame:
     data[_dcfg.real_unknown] = data[_dcfg.real_unknown].interpolate(method='linear', limit=2).ffill().bfill()
     _warn_missing(data, _dcfg.real_known, 'daily')
     data[_dcfg.real_known]   = data[_dcfg.real_known].interpolate(method='linear', limit=2).ffill(limit=1)
-    data[_dcfg.targets]      = data[_dcfg.targets].fillna(0)
+    data[_dcfg.targets]      = data[_dcfg.targets].fillna(0).astype(float)
     data.reset_index(drop=True, inplace=True)
     data['static']   = 'S'
     data['time_idx'] = np.arange(data.shape[0])
